moved some validator utils to tendermint utils, election approval now handled by planetmint
